+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/SpecConfig.java
@@ -138,7 +138,7 @@ public interface SpecConfig {
 
   long getDepositChainId();
 
-  int getDepositNetworkId();
+  long getDepositNetworkId();
 
   Eth1Address getDepositContractAddress();
